Move plugin integration into own class
